+++ src/Molecules/BottomSheet/BottomSheet.tsx
@@ -1,35 +1,50 @@
-import React, { useRef } from 'react';
+import React, { useRef, useState } from 'react';
 import FocusLock from 'react-focus-lock';
 import { AnimatePresence, motion } from 'framer-motion';
 import { RemoveScroll } from 'react-remove-scroll';
 
-import { Button, Flexbox, Icon, Portal, theme, useOnClickOutside } from '../..';
-import { Backdrop, StyledBottomSheet } from './BottomSheet.styles';
+import { Flexbox, Icon, Portal, theme, useOnClickOutside } from '../..';
+import { Backdrop, StyledBottomSheet, StyledIconButton } from './BottomSheet.styles';
+import { isBoolean } from '../../common/utils';
 import { Props } from './BottomSheet.types';
+import { TRANSITION_DURATION } from './constants';
 
 const BottomSheet: React.FC<Props> = ({
   children,
   className,
   closeOnClickOutside,
   fullScreenMobile,
-  height = '40%',
-  onClose,
-  open,
+  height,
+  invertedColors,
+  onClose: onCloseExternal,
+  open: isOpenExternal,
   showBackdrop = true,
   title,
 }) => {
   const internalRef = useRef<HTMLDivElement>(null);
+  const isControlled = isBoolean(isOpenExternal);
+  const [isOpenInternal, setIsOpenInternal] = useState(true);
+
+  const onClose = () => {
+    setIsOpenInternal(false);
+
+    if (typeof onCloseExternal === 'function') {
+      onCloseExternal();
+    }
+  };
 
   useOnClickOutside(internalRef, () => {
-    if (closeOnClickOutside && typeof onClose === 'function') {
+    if (closeOnClickOutside) {
       onClose();
     }
   });
 
+  const shouldRender = isControlled ? isOpenExternal : isOpenInternal;
+
   return (
     <Portal>
       <AnimatePresence>
-        {open ? (
+        {shouldRender ? (
           <FocusLock disabled={fullScreenMobile}>
             <RemoveScroll enabled={!fullScreenMobile}>
               {showBackdrop && (
@@ -38,7 +53,7 @@ const BottomSheet: React.FC<Props> = ({
                   initial={{ opacity: 0, zIndex: theme.zIndex.overlay, position: 'fixed' }}
                   exit={{ opacity: 0 }}
                   animate={{ opacity: 1 }}
-                  transition={{ type: 'ease', duration: 0.2 }}
+                  transition={{ type: 'ease', duration: TRANSITION_DURATION }}
                 >
                   <Backdrop container alignItems="center" justifyContent="center" />
                 </motion.div>
@@ -52,23 +67,29 @@ const BottomSheet: React.FC<Props> = ({
                 }}
                 exit={{ bottom: '-100%' }}
                 animate={{ bottom: 0 }}
-                transition={{ type: 'ease', duration: 0.2 }}
+                transition={{ type: 'ease', duration: TRANSITION_DURATION }}
                 style={{ width: '100%', height: '100%' }}
               >
                 <StyledBottomSheet
                   className={className}
-                  fullScreenMobile={fullScreenMobile}
+                  $fullScreenMobile={fullScreenMobile}
                   height={height}
+                  $invertedColors={invertedColors}
                   ref={internalRef}
                 >
                   <Flexbox container direction="column" gap={2}>
                     <Flexbox container item justifyContent="space-between" alignItems="center">
                       <Flexbox item>{title}</Flexbox>
                       {typeof onClose === 'function' && (
-                        <Flexbox item>
-                          <Button.Icon variant="primary" onClick={onClose} size="s">
-                            <Icon.Cross16 color="currentColor" />
-                          </Button.Icon>
+                        <Flexbox item alignSelf="flex-start">
+                          <StyledIconButton
+                            variant="primary"
+                            onClick={onClose}
+                            size="s"
+                            $invertedColors={invertedColors}
+                          >
+                            <Icon.Cross16 />
+                          </StyledIconButton>
                         </Flexbox>
                       )}
                     </Flexbox>