+++ src/Molecules/FilterChip/FilterChip.styled.ts
@@ -7,36 +7,39 @@ const variantStyles = (theme: Theme, variant: string, $selected: boolean) =>
     default: css`
       ${$selected &&
       `
-      color: ${theme.color.quickFilterSelectedText};
-      background: ${theme.color.quickFilterSelectedBackground};
+      color: ${theme.colorTokens.action.text_default};
+      background: ${theme.colorTokens.action.background_weak};
+      &:hover {
+        color: ${theme.colorTokens.action.text_hover};
+      }
       &:focus-visible {
-        outline: 1px solid ${theme.color.quickFilterFocusSelectedOutline};
+        outline: 1px solid ${theme.colorTokens.action.border_default};
       }
       `}
     `,
     sustainability: css`
       &:hover {
-        color: ${theme.color.quickFilterSustainabilityHoverColor};
+        color: ${theme.colorTokens.positive.text_hover};
       }
       ${$selected &&
       `
-      color: ${theme.color.quickFilterSustainabilityColor};
-      background: ${theme.color.quickFilterSustainabilityBackground};
+      color: ${theme.colorTokens.positive.text_default};
+      background: ${theme.colorTokens.positive.background_weak};
       &:focus-visible {
-        outline: 1px solid ${theme.color.quickFilterSustainabilityColor};
+        outline: 1px solid ${theme.colorTokens.positive.border_default};
       }
       `}
     `,
     brand: css`
       ${$selected &&
       `
-      color: ${theme.color.quickFilterBrandActiveColor};
-      background: ${theme.color.quickFilterBrandActiveBackground};
+      color: ${theme.colorTokens.neutral.text_strong};
+      background: ${theme.colorTokens.neutral.background_brand_black};
       &:focus-visible {
-        outline: 1px solid ${theme.color.quickFilterBrandFocusBorder};
+        outline: 1px solid ${theme.colorTokens.action.border_default};
       }
       &:hover {
-        color: ${theme.color.quickFilterBrandHoverColor};
+        color: ${theme.colorTokens.action.text_default};
       }
       `}
     `,
@@ -55,16 +58,16 @@ export const ContainerLabel = styled.label<{
   cursor: pointer;
   padding: ${(p) => p.theme.spacing.unit(2)}px;
   height: ${(p) => p.theme.spacing.unit(8)}px;
-  color: ${(p) => p.theme.color.quickFilterText};
-  background: ${(p) => p.theme.color.quickFilterBackground};
+  color: ${(p) => p.theme.colorTokens.neutral.text_default};
+  background: ${(p) => p.theme.colorTokens.neutral.background_medium};
   outline-offset: -1px;
 
   &:hover {
-    color: ${(p) => p.theme.color.quickFilterSelectedText};
+    color: ${(p) => p.theme.colorTokens.action.text_hover};
   }
 
   &:focus-visible {
-    outline: 1px solid ${(p) => p.theme.color.quickFilterFocusOutline};
+    outline: 1px solid ${(p) => p.theme.colorTokens.neutral.border_focus};
   }
 
   transition:
@@ -88,14 +91,14 @@ export const ContainerLabel = styled.label<{
   ${(p) =>
     p.$disabled &&
     `
-      background: ${p.theme.color.disabledBackground};
-      color: ${p.theme.color.disabledText};
+      background: ${p.theme.colorTokens.neutral.background_disabled};
+      color: ${p.theme.colorTokens.neutral.text_disabled};
       cursor: not-allowed;
       &:focus-visible {
         outline: none;
       }
       &:hover {
-        color: ${p.theme.color.disabledText};
+        color: ${p.theme.colorTokens.neutral.text_disabled};
       }
     `};
 `;